remove subscribeOn from service and let usecase use Schedulers.io() thread by default
