+++ DriverPortal/app/src/main/java/com/edulog/driverportal/common/domain/UseCase.java
@@ -4,11 +4,14 @@ import io.reactivex.Observable;
 import io.reactivex.Observer;
 import io.reactivex.Scheduler;
 import io.reactivex.android.schedulers.AndroidSchedulers;
+import io.reactivex.schedulers.Schedulers;
 
 public abstract class UseCase<T, Params> {
+    private Scheduler jobExecutionScheduler;
     private Scheduler postExecutionScheduler;
 
     public UseCase() {
+        this.jobExecutionScheduler = Schedulers.io();
         this.postExecutionScheduler = AndroidSchedulers.mainThread();
     }
 
@@ -18,11 +21,16 @@ public abstract class UseCase<T, Params> {
         Observable<T> observable = buildUseCaseObservable(params);
 
         observable
+                .subscribeOn(jobExecutionScheduler)
                 .observeOn(postExecutionScheduler)
                 .subscribe(observer);
     }
 
     public void setPostExecutionScheduler(Scheduler scheduler) {
         this.postExecutionScheduler = scheduler;
+    }
+
+    public void setJobExecutionScheduler(Scheduler scheduler) {
+        this.jobExecutionScheduler = scheduler;
     }
 }
