+++ DriverPortal/app/src/main/java/com/edulog/driverportal/routedetails/MapService.java
@@ -6,7 +6,7 @@ import retrofit2.http.Query;
 
 public interface MapService {
     @GET("/maps/api/directions/json")
-    Observable<EncodedPolylineEntity> getDirection(@Query("origin") String origin,
-                                                   @Query("destination") String destination,
-                                                   @Query("key") String key);
+    Observable<PolylineEntity> getDirection(@Query("origin") String origin,
+                                            @Query("destination") String destination,
+                                            @Query("key") String key);
 }
