--- DriverPortal/app/src/main/java/com/edulog/driverportal/routes/data/net/DriverPortalRouteService.java
@@ -1,18 +1,0 @@
-package com.edulog.driverportal.routes.data.net;
-
-import com.edulog.driverportal.routes.data.entity.RouteEntity;
-
-import java.util.List;
-
-import io.reactivex.Observable;
-import retrofit2.http.GET;
-import retrofit2.http.Path;
-import retrofit2.http.Query;
-
-public interface DriverPortalRouteService {
-    @GET("/api/routes/{route_id}")
-    Observable<RouteEntity> getRoute(@Path("route_id") String routeId);
-
-    @GET("/api/search/routes")
-    Observable<List<RouteEntity>> findRoutes(@Query("route_id") String query);
-}
