+++ DriverPortal/app/src/main/java/com/edulog/driverportal/routes/domain/service/RouteService.java
@@ -5,9 +5,14 @@ import com.edulog.driverportal.routes.data.entity.RouteEntity;
 import java.util.List;
 
 import io.reactivex.Observable;
+import retrofit2.http.GET;
+import retrofit2.http.Path;
+import retrofit2.http.Query;
 
 public interface RouteService {
-    Observable<RouteEntity> getRoute(String routeId);
+    @GET("/api/routes/{route_id}")
+    Observable<RouteEntity> getRoute(@Path("route_id") String routeId);
 
-    Observable<List<RouteEntity>> findRoutes(String query);
+    @GET("/api/search/routes")
+    Observable<List<RouteEntity>> findRoutes(@Query("route_id") String query);
 }
