+++ DriverPortal/app/src/main/java/com/edulog/driverportal/routes/presentation/view/fragment/PreviewRouteDialogFragment.java
@@ -13,13 +13,11 @@ import android.widget.TextView;
 import com.edulog.driverportal.R;
 import com.edulog.driverportal.common.presentation.BaseActivity;
 import com.edulog.driverportal.util.RetrofitServiceGenerator;
-import com.edulog.driverportal.routes.data.net.DriverPortalRouteService;
+import com.edulog.driverportal.routes.domain.service.RouteService;
 import com.edulog.driverportal.routes.data.repository.DriverPortalDbHelper;
 import com.edulog.driverportal.routes.data.repository.RouteRepositoryImpl;
-import com.edulog.driverportal.routes.data.service.RouteServiceImpl;
 import com.edulog.driverportal.routes.domain.interactor.GetRouteUseCase;
 import com.edulog.driverportal.routes.domain.repository.RouteRepository;
-import com.edulog.driverportal.routes.domain.service.RouteService;
 import com.edulog.driverportal.routes.model.LoadMode;
 import com.edulog.driverportal.routes.model.RouteModel;
 import com.edulog.driverportal.routes.presentation.presenter.PreviewRouteContract;
@@ -56,8 +54,7 @@ public class PreviewRouteDialogFragment extends DialogFragment implements Previe
             loadMode = (LoadMode) getArguments().getSerializable(KEY_LOAD_MODE);
         }
 
-        DriverPortalRouteService service = new RetrofitServiceGenerator().generate(DriverPortalRouteService.class);
-        RouteService routeService = new RouteServiceImpl(service);
+        RouteService routeService = new RetrofitServiceGenerator().generate(RouteService.class);
         RouteRepository routeRepository = new RouteRepositoryImpl(new DriverPortalDbHelper(getActivity()));
         GetRouteUseCase getRouteUseCase = new GetRouteUseCase(routeService, routeRepository);
         previewRoutePresenter = new PreviewRoutePresenterImpl(getRouteUseCase);
