+++ DriverPortal/app/src/main/java/com/edulog/driverportal/settings/changepassword/data/AuthServiceImpl.java
@@ -1,11 +1,10 @@
-package com.edulog.driverportal.settings.changepassword.data.service;
+package com.edulog.driverportal.settings.changepassword.data;
 
 import com.edulog.driverportal.util.RetrofitServiceGenerator;
-import com.edulog.driverportal.settings.changepassword.data.net.ChangePasswordService;
+import com.edulog.driverportal.settings.changepassword.domain.service.ChangePasswordService;
 import com.edulog.driverportal.settings.changepassword.domain.service.AuthService;
 
 import io.reactivex.Observable;
-import io.reactivex.schedulers.Schedulers;
 import okhttp3.ResponseBody;
 import retrofit2.Response;
 
@@ -15,7 +14,6 @@ public class AuthServiceImpl implements AuthService {
         ChangePasswordService service = new RetrofitServiceGenerator().generate(ChangePasswordService.class);
         Observable<Response<ResponseBody>> observable = service.changePassword(driverId, oldPassword, newPassword);
         return observable
-                .subscribeOn(Schedulers.io())
                 .map(response -> response.code() == 200);
     }
 }
