+++ DriverPortal/app/src/test/java/com/edulog/driverportal/routes/domain/interactor/GetRouteUseCaseTest.java
@@ -3,7 +3,6 @@ package com.edulog.driverportal.routes.domain.interactor;
 import com.edulog.driverportal.RxImmediateSchedulerRule;
 import com.edulog.driverportal.routes.data.entity.RouteEntity;
 import com.edulog.driverportal.routes.domain.repository.RouteRepository;
-import com.edulog.driverportal.routes.domain.service.RouteService;
 import com.edulog.driverportal.routes.model.LoadMode;
 
 import org.junit.Before;
