+++ DriverPortal/app/src/test/java/com/edulog/driverportal/routes/domain/interactor/SearchRoutesUseCaseTest.java
@@ -1,7 +1,6 @@
 package com.edulog.driverportal.routes.domain.interactor;
 
 import com.edulog.driverportal.RxImmediateSchedulerRule;
-import com.edulog.driverportal.routes.domain.service.RouteService;
 
 import org.junit.Before;
 import org.junit.ClassRule;
